backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
python-docx==1.1.2
reportlab==4.2.2
